Add RunServer() and TestServer() functions + add Config struct
There are now two new functions for running a gotables server and testing the connection. Additionally, the information for the server and the connection are now in a config struct. Note: The RunServer() function can only run servers locally. Setting the Host field in Config to something external will result in failure when testing the connection if the server was started using this module.